# swagger.yaml
openapi: 3.0.4
info:
  title: Clone X API
  description: |-
    API documentation for Clone X - A Twitter-like social media platform.

    This API provides endpoints for authentication, user management, tweets, and messaging functionality.
  termsOfService: https://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://your-production-url.com
    description: Production server
...
paths:
...
  /api/user/{followed_user_id}/follow:
    post:
      tags: [user]
      summary: Follow a user
      description: Atomically creates the follower relation and updates both counters. With the follow-outbox migration enabled, notification delivery is asynchronous and rechecks the relation and two-way blocks.
      operationId: followUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FollowedUserId"
      responses:
        "200":
          description: Follow relation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NullDataSuccessResponse"
        "400":
          description: Invalid user ID or self-follow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Follow is forbidden by a block in either direction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Target user does not exist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Follow relation already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [user]
      summary: Unfollow a user
      description: Atomically removes the follower relation and decrements both counters. Historical notifications are retained; a pending follow event is suppressed by worker-time relation validation.
      operationId: unfollowUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FollowedUserId"
      responses:
        "200":
          description: Follow relation removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NullDataSuccessResponse"
        "400":
          description: Invalid user ID or follow relation does not exist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Target user does not exist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/user/{followed_user_id}/follow-notification-preferences:
    patch:
      tags: [user]
      summary: Update followed-account post notification preference
      description: Updates only the authenticated user's existing follow relation. New relations default to posts=false and unfollow removes the preference with the relation.
      operationId: updateFollowNotificationPreference
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FollowedUserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [posts]
              properties:
                posts:
                  type: boolean
      responses:
        "200":
          description: Preference updated
          content:
            application/json:
              schema:
                type: object
                required: [statusCode, message, data]
                properties:
                  statusCode:
                    type: integer
                    example: 200
                  message:
                    type: string
                  data:
                    type: object
                    required: [followed_user_id, posts]
                    properties:
                      followed_user_id:
                        type: string
                        pattern: "^[a-fA-F0-9]{24}$"
                      posts:
                        type: boolean
        "400":
          description: Invalid user ID or body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: FOLLOW_RELATION_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/notifications:
    get:
      tags:
        - notification
      summary: List notifications
      description: Returns non-invalidated notifications ordered by immutable tuple created_at descending then _id descending. next_cursor is opaque; a legacy 24-hex ObjectId cursor is still accepted when it identifies a notification owned by the caller. Pagination fetches one extra record so has_next_page is exact. During asynchronous lifecycle cleanup, an actor or target that is no longer visible is redacted at read time; its raw sender_id/target_id becomes null, actor_ids_preview is filtered, context is empty, and deduplication/aggregation keys are omitted.
      operationId: getNotifications
      security:
        - bearerAuth: []
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: Use next_cursor verbatim. Opaque v2 cursors and legacy 24-hex ObjectId cursors are accepted.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        "200":
          description: Notification page returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPageResponse"
        "400":
          description: Invalid cursor or limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
...
    FollowedUserId:
      name: followed_user_id
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-fA-F0-9]{24}$"
...
  schemas:
    NullDataSuccessResponse:
      type: object
      required: [statusCode, message, data]
      properties:
        statusCode:
          type: integer
          example: 200
        message:
          type: string
        data:
          nullable: true
          example: null
...
    LegacyNotificationFields:
      type: object
      required:
        - _id
        - recipient_id
        - sender_id
        - type
        - target_id
        - is_read
        - created_at
      properties:
        _id:
          type: string
          pattern: "^[a-fA-F0-9]{24}$"
        recipient_id:
          type: string
          pattern: "^[a-fA-F0-9]{24}$"
        sender_id:
          type: string
          nullable: true
          pattern: "^[a-fA-F0-9]{24}$"
        type:
          type: string
          enum: [like, reply, retweet, quote, follow, mention, system, message, message_reply, message_mention, message_reaction, group_add, group_join, group_kick, admin_granted, admin_revoked, followed_user_tweet]
        target_id:
          type: string
          nullable: true
          pattern: "^[a-fA-F0-9]{24}$"
        is_read:
          type: boolean
        created_at:
          type: string
          format: date-time
    NotificationActorInfo:
      type: object
      required: [_id, name, username]
      properties:
        _id:
          type: string
          pattern: "^[a-fA-F0-9]{24}$"
        name:
          type: string
        username:
          type: string
        avatar:
          type: string
    NotificationTargetInfo:
      type: object
      required: [_id, target_type]
      properties:
        _id:
          type: string
          pattern: "^[a-fA-F0-9]{24}$"
        target_type:
          type: string
          enum: [USER, TWEET, MESSAGE, CONVERSATION]
        name:
          type: string
        username:
          type: string
        avatar:
          type: string
        owner_id:
          type: string
          pattern: "^[a-fA-F0-9]{24}$"
        tweet_type:
          type: integer
          enum: [0, 1, 2, 3]
        content:
          type: string
          maxLength: 140
        conversation_id:
          type: string
          pattern: "^[a-fA-F0-9]{24}$"
        sender_id:
          type: string
          pattern: "^[a-fA-F0-9]{24}$"
        status:
          type: string
          enum: [sent, revoked, deleted]
        conversation_type:
          type: string
          enum: [direct, group]
        avatar_url:
          type: string
    NotificationV2Fields:
      type: object
      properties:
        target_type:
          type: string
          nullable: true
          enum: [USER, TWEET, MESSAGE, CONVERSATION]
        actor_ids_preview:
          type: array
          maxItems: 3
          description: Contains only actor IDs currently visible to the recipient; hidden actors are removed rather than represented as null entries.
          items:
            type: string
            pattern: "^[a-fA-F0-9]{24}$"
        actor_count:
          type: integer
          minimum: 0
        context:
          type: object
          additionalProperties: true
        deduplication_key:
          type: string
          description: Omitted from REST list items when the actor or target is redacted.
        aggregation_key:
          type: string
          description: Omitted from REST list items when the actor or target is redacted.
        aggregation_active:
          type: boolean
        read_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
        invalidated_at:
          type: string
          format: date-time
          nullable: true
        schema_version:
          type: integer
          enum: [2]
...
    NotificationListItem:
      allOf:
        - $ref: "#/components/schemas/LegacyNotificationFields"
        - $ref: "#/components/schemas/NotificationV2Fields"
        - type: object
          required:
            - target_type
            - actor_ids_preview
            - actor_count
            - context
            - aggregation_active
            - read_at
            - updated_at
            - invalidated_at
            - actor_info
            - actor_infos_preview
            - target_info
          properties:
            actor_info:
              nullable: true
              allOf:
                - $ref: "#/components/schemas/NotificationActorInfo"
            actor_infos_preview:
              type: array
              maxItems: 3
              items:
                nullable: true
                allOf:
                  - $ref: "#/components/schemas/NotificationActorInfo"
            target_info:
              nullable: true
              allOf:
                - $ref: "#/components/schemas/NotificationTargetInfo"
    NotificationPageData:
      type: object
      required:
        - notifications
        - unreadCount
        - next_cursor
        - has_next_page
      properties:
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/NotificationListItem"
        unreadCount:
          type: integer
          minimum: 0
        next_cursor:
          type: string
          nullable: true
          description: Opaque pagination cursor. Send it back verbatim; do not parse it.
        has_next_page:
          type: boolean
    NotificationPageResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          required:
            - data
          properties:
            data:
              $ref: "#/components/schemas/NotificationPageData"
...
    SuccessResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 200
        message:
          type: string
          example: "Operation successful"
      required:
        - statusCode
        - message
...
    Error:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
          description: HTTP status code
        message:
          type: string
          example: "Validation failed"
          description: Error message describing what went wrong
        error:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            "body.email": ["Email must be valid"]
            "body.password": ["Password is required"]
          description: Detailed validation errors (optional, only present for validation failures)
        code:
          type: string
          description: Stable optional application error code; clients should branch on this value instead of parsing message text.
          example: "GROUP_SOLE_ADMIN_CANNOT_LEAVE"
      required:
        - statusCode
        - message
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT